Create Cleaning attempt.R
I tried to clean a bit our data. I used Leslie's function and a part of Carla's code. 
I also discovered that in the applications dataset we have some ProjectsID that are not in the reviewers dataset. Probably that's the reason why when we merge the data we loose some observations.