Update
As the regression data has now Proposal combined, I took this variable out from this regressions.

For completeness I added plots for ScientificRelevance and Suitability to visualize the perfect separation.

To be consequent all over, I added at the begining Interaction between Institution and Division.

Add some plots for the effect of gender in the different grades.